Bug: Fixed bug where youtube gallerites won't play properly.
A user logged in to YouTube can open a video and continue watching where they left the video.
In this case the URL wil have a time parameter at the end '&t=[time]s'. This paramter breaks the embed link when diplayed on the page.
So a line is added to make sure it is removed from the link.