Bug: Fixed bug with editing YouTube link.
If the video has time parameter things worked fine because the index of slice was found properly.
But if it was a new video, index of '&t=' would be -1 which leads to improper slicing 'removing last character in link'.
So an if statement now does the slice properly if the time parameter is not in the link.